requirements.txt
# AWS SDK
boto3==1.34.131
botocore==1.34.131

# Strands Agents (do not pin)
strands-agents

# Web framework (optional)
Flask==3.0.2

# JSON + data utilities
pydantic==2.6.4
python-dotenv==1.0.1

# Networking + progress
requests==2.31.0
tqdm==4.66.1

# Dev helper
watchdog==4.0.0

# Telemetry (stable version for strands)
opentelemetry-sdk==1.22.0

openai

strands-agents[ollama]